requirements.txt
fastapi
uvicorn
pytesseract
Pillow
transformers
sentence-transformers
chromadb
streamlit
requests